Box Rust code's context and pass that into the API
This ensure's that the point we're pass around FFI boundaries is
probably aligned and valid.

I think this means we're trusting libsettings to not do anything bad
with multiple mutable Context values floating around.

When creating the context, instead of capturing the context
itself we're capture the pointer to the context, which will allow
us to dig out a mutable reference when the callback is
invoked.

Again this assumes that libsettings isn't doing anything weird
with the mutable reference.