Initial implementation of ObservationTableModel
This is a first take at reimplementing the model used by the observation
tab as a QAbstractTableModel, serving up individual cell updates using
the API defined in QAbstractItemModel.

The UI changes I've made for this aren't quite working yet.
At present, the data appears to be getting up to QML properly, but for
some reason the table itself is not being rendered.

A benefit this has over the other one is that only values that change
will be updated in the UI. The self.dataChanged signal is the magic
that will let QML know when only certain columns need updating.

week and tow properties I kind of rolled into the Table Model. It's
certainly not required for the table model, but it is data that is
related to the OBS tables. I could have split them out separately, but a
separate class registered with just those two things seemed silly.
Still, it probably should be separated.

The notify signal for those properties will allow the QML binding to
these properties update the UI whenever notification of change occurs -
that wasn't in play in your implementation. I saw use of properties, but
not of notify signals.

Without the notify signals, the QML bindings don't know when to update
the UI with the updated value - it doesn't know when the value was
updated at all.

QML still controls the update frequency with the timer in
ObservationTab, but what is done in that timer handler is different.
Instead of asking python to fill a data structure, and then copying that
on to javascript/qml variables in the QML implemented table model,
instead just an update method is invoked on each of the local and remote
tables, which in turn calls update on the python model implementation,
which will do the work of going through the REMOTE_OBSERVATION_TAB or
LOCAL_OBSERVATION_TAB data structures that are updated with the incoming
data from rust, and comparing them with the data values in the copy that
is contained in the ObservationTableModel instance. If it finds a
change in a data value, it will update the instance's copy, and signal
QML with the dataChanged signal, indicating the row and column that
changed. QML then will only redraw the particular cell that changed (or
not at all if it's not in view).

I also changed the logic for setting visibility of items, so it is
controlled from the TableView itself.