Add organization name, domain and app name to QApplication
FileDialog uses QSettings, which really wants applications to have
an organizationName, organizationDomain, and applicationName set so that
settings can be saved to a namespace that is unique to the application.
This sets that.

Pass arguments to QApplication, so platform plugins and other features
can be specified on the console_pp command line for Qt to act on.