ObservationTab performance optimizations (#125)
* Update readme to inform users to deactivate conda

* Inform devs to deactivate conda if they have activated the console_pp
  conda environment prior to installing or running the console_pp
  project. The environment is explicitly selected when you invoke these
  commands.

* Initial implementation of ObservationTableModel

This is a first take at reimplementing the model used by the observation
tab as a QAbstractTableModel, serving up individual cell updates using
the API defined in QAbstractItemModel.

The UI changes I've made for this aren't quite working yet.
At present, the data appears to be getting up to QML properly, but for
some reason the table itself is not being rendered.

A benefit this has over the other one is that only values that change
will be updated in the UI. The self.dataChanged signal is the magic
that will let QML know when only certain columns need updating.

week and tow properties I kind of rolled into the Table Model. It's
certainly not required for the table model, but it is data that is
related to the OBS tables. I could have split them out separately, but a
separate class registered with just those two things seemed silly.
Still, it probably should be separated.

The notify signal for those properties will allow the QML binding to
these properties update the UI whenever notification of change occurs -
that wasn't in play in your implementation. I saw use of properties, but
not of notify signals.

Without the notify signals, the QML bindings don't know when to update
the UI with the updated value - it doesn't know when the value was
updated at all.

QML still controls the update frequency with the timer in
ObservationTab, but what is done in that timer handler is different.
Instead of asking python to fill a data structure, and then copying that
on to javascript/qml variables in the QML implemented table model,
instead just an update method is invoked on each of the local and remote
tables, which in turn calls update on the python model implementation,
which will do the work of going through the REMOTE_OBSERVATION_TAB or
LOCAL_OBSERVATION_TAB data structures that are updated with the incoming
data from rust, and comparing them with the data values in the copy that
is contained in the ObservationTableModel instance. If it finds a
change in a data value, it will update the instance's copy, and signal
QML with the dataChanged signal, indicating the row and column that
changed. QML then will only redraw the particular cell that changed (or
not at all if it's not in view).

I also changed the logic for setting visibility of items, so it is
controlled from the TableView itself.

* ObservationTable QAbstractTableModel finished impl

* This now has a complete working QAbstractTableModel implementation for
  the ObservationTable. The table displays now, and updates are
  delivered on a per-cell change basis, so only those cells that have
  data that has changed are redrawn, versus the redraw of the whole
  table on each timer fire that was occurring with the prior implementation.

* The missing piece was that when rows are added or removed to/from a
  QAbstractItemModel, QAbstractTableModel, or QAbstractListModel after
  creation, those row additions or subtractions need to be announced,
  causing signals to be sent to the UI, so the UI knows which parts of
  the table to update, or create delegates for.

* Cleaned up access to the ObservationTableModel on the QML side,
  improving where it is constructed and accessed.

* Week and tow are pulled from the table model now.

* ObservationData and ObservationModel types are now unused and removed.
  Everything contained in both of those is now in the
  ObservationTableModel.

* Remove unused observation_tab python debug code

* Remove unneeded elements in SolutionTab, Table

* This removes unnecessary Rectangle in SolutionTab

* This removes unnecessary parent width and height specifiers in
  SolutionTable. referencing parent elements in a toplevel QML element
  is not recommended. Instead, specify the sizing from wherever you use
  the QML component you have defined. If you need a fixed size for the
  element, then you can set it in the defined toplevel component.

* Remove extraneous width, height set when anchors.fill was already
  used.

* Optimize SolutionTable syncColumnWidthsWithSplitView

* This eliminates passing in the oldcols into
  syncColumnWidthsWithSplitView, and instead does that within the
  function.
* Omit calling syncColumnWidthsWithSplitView on every timer fire.
  Instead, perform it when the width of the TableView changes. This will
  change whenever the splitter is moved, and upon first display.

* Fix code quality warnings

This just fixes pylint warnings and formatting issues.

* Fix and ignore mypy PySide interaction issues

* Fix one issue that mypy reports that is, technically accurate, but
  could be fixed by doing it the way mypy indicated was necessary
  (Specifying a single argument to a signal by specifying
   arguments=<str> instead of arguments=<List of str>).

  Both forms are valid (second one is mandatory if you are defining a
  signal that takes more than one argument), but mypy doesn't recognize
  the second form as valid. Since these signals only take a single
  argument, the form that mypy is happy with can be used.

* Signal emit is something that PySide does not handle right now in it's
  pyi files that mypy uses. The following link discusses it:
  https://bugreports.qt.io/browse/PYSIDE-1638?focusedCommentId=581411&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-581411
  Ignore this mypy warning.

* Slot not being callable also seems to be an issue with PySide's .pyi
  files, so we have mypy ignore these. They are perfectly valid.
  Ignore this mypy warning.

* Property notify not taking a Signal is also valid, but mypy doesn't
  recognize it as valid. I believe this is also a PySide2 .pyi issue -
  though I don't have any docs to back that up. It is a Signal, which
  would suggest it's a known issue.
  Ignore this mypy warning.

* rowData slot not needed anymore - it was used for debug, so removed.