TrackingSignalsTab - don't recreate series
This is the first iteration of optimizing performance of the
TrackingSignalsTab.

Series' are created only once in QML, to associate them with the chart.
Thereafter, the series' persist. Points are replaced and properties are
set with each fill_all_series call - which occurs regularly as called by
a timer from QML. The update mechanism remains unchanged.

Legend seems broken with the changes I have put in here - despite my
having attempted to accomodate it.