[CPP-347] Improve handling of firmware versions (#174)
* [CPP-347] Improve handling of firmware versions

Introduces a new 'SwiftVersion' class that can be used for comparing
firmware and console versions. This was heavily inspired/ported from the
old console and removes the need for the SemVer package.

As part of this, firmware upgrades from before v2.0.0 to after v2.0.0
are not allowed, as in the old console.