workaround for linking bug in pyo3
The pyo3 build script removes python library link directives if you're
building an extension module.

This is problem if you want to build a binary that links against Python
and an extension module in the same project (like we are).

Rust features are additive, so it leaves the "extension-module" feature
enabled in the "shared" build of pyo3, this causes pyo3 to fail to emit
linking directives when building the console binary.

This only worked before because we were using two different versions of
pyo3.