Windows installer redist support.[CPP-324] (#237)
First checks the registry to see if VS C++ 14 x64 is present. If not it will attempt to download and install. Then does another check to see if the registry is now valid. If not it will create a popup:
Seems UAC is needed to install this redistributable installer but in theory should only be needed once as future updates will have it already installed. So the auto updater idea is still viable.