Create base `Tab` type
* This creates a base qml `Tab` type that will be used for all main tabs
  within the application.
* Create properties within Tab for the sub-tab names that the tab will
  have (if any), and which sub-tab is the current one.

This is prep work to allow the sub-tabs of a tab to be held and
presented from a parent item of the current tab - having a common TabBar
showing sub-tabs for all main tabs, instead of having a TabBar within
each tab that has sub-tabs.

This is added to support the new design guidelines that have the TabBar
showing sub-tabs extend outside of the region that a Tab is contained
within (extending to the left, above the SideNavBar).