Increase cases under which TabInfoBar autohides
Now if the mouse moves off the TabInfoBar, and the TabInfoBar has
subtabs, it will auto-hide after delay.

Pressing the Open button will allow the TabInfoBar to remain open, as
long as the mouse pointer does not enter/exit the TabInfoBar (so, to get
this sticky behavior, you need to move the mouse down fairly quickly).