Prevent table rows from drawing over header
Prevent the SolutionTable rows from drawing over the table header.
They were drawing outside of the boundary of their item. The solution is
to turn on clipping.

Clipping is not enabled by default because it does incur some
performance penalty, however in this case it is pretty negligible.

I also removed clipping for the items in the table header, as there was
no reason to clip those - they do not draw outside of their boundaries.