Override Controls 2 Label for explicit font size
This overrides Qt Quick Control 2's Label so an explicit font size can
be set to it, as it did not seem to be honoring the qtquickcontrols
default font size.

This also sets the default font in the qtquickcontrols2.conf file to
Roboto Condensed to match the setting we made in Constants.fontFamily.