fileio: Fix timeout on serial write [CPP-644]
Fixes a condition where the loop that's waiting for
MSG_FILEIO_WRITE_RESP messages can mistakenly think that the write has
concluded and exit too early. This was causing the fileio program to
hang as the other threads never terminated.

From what I can tell this was happening because there was a bit of a
race between that thread checking that there were no pending writes and
the `req_maker_thread` inserting new pending writes. I think this might
have been exacerbated on serial because the `send` call takes longer,
which gives a greater chance of the waiting loop to see no pending
writes and means that there can potentially be less "in flight" messages
at once.