Add lldb python api use for iterating
This adds a means to loop rerun the console debugging session after a
normal exit from lldb - speeding up each run session by cutting out the
unload and load of the debugger and the executable.

Note: this still has issues with the echoing of the lldb REPL. I have
not solved that yet.