Properly check for QTDIR env var
Duckscript does not use environment variables as regular variables,
except when entering a new duckscript script - then duckscript variables
of the same name are populated with environment variable values. So
if you set an environment variable, you must fetch it with get_env to be
able to check against it.