Fix QML Layouts recursive rearrange error
This fixes an error in laying out buttons in the connection dialog.
References to the parent layout's width while defining the button widths
is recursive - as the preferred widths of the buttons would affect the
the width of the parent, so referencing the parent width to determine
the width of the buttons is no good.
Replace instead with referencing the width of the parent's width - as
the parent layout fills its parent.

An alternative fix that was not chosen was to instead define the
minimumWidth of the buttons instead of the preferredWidth.