LoggingBar - Use QStringListModel for some properties
LoggingBarData included a couple of properties that were passing lists
of strings from Python to QML, and storing these in javascript arrays.

To improve efficiency, and what was an attempt to migrate away from
using QVariantLists across the Python QML boundary, the QVariantLists
were abandoned here, and the properties were exposed using
QStringListModels. Here I derived a new custom type from them, just so I
could provide a convenience method to return the index of a particular
string in the model. As these are QAbstractItemModels, they can be used
directly as the model property in the QComboBoxes that show them.