Remove use of MetricsMonitor in the app
MetricsMonitor is never enabled, and is just there causing QML warning messages
for no reason. The warnings could be fixed, but there is no way to test them at
present, so, for now, just disable use of it in the app. Leave the
MetricsMonitor.qml present, but don't package it with the app (leave it out of
console_resources.qrc).

When it is to be re-enabled, then it can be fixed up and re-tested.