Add Qt releasing swift-nav 5.15.2.2 wheels with recursive mutex locking
Qt Releasing was able to cut a special build of PySide2 5.15.2.1 with
the final-2 patch that should address the occasional crashes in PySide2
signalmanager, as well as the hangs on shutdown that Jason was
experiencing at the end of last week.

The fixes (targetting the dev branch) can be seen here (these are publicly viewable):
    https://codereview.qt-project.org/c/pyside/pyside-setup/+/413917
    https://codereview.qt-project.org/c/pyside/pyside-setup/+/415523

The fixes ported to the 5.15 branch are not public, but if one has a
commercial license assigned to their Qt account, you can see them. They
are here:
    https://codereview.qt-project.org/c/pyside/tqtc-pyside-setup/+/414778
    https://codereview.qt-project.org/c/pyside/tqtc-pyside-setup/+/415707

The change has been cherry-picked to 5.15 LTS, 6.2 LTS, and 6.3
branches, so the next (from today, 20220613) official Qt releases of
each of these Qt versions will include the fix. For now, please use
these wheels.

They can be used by performing the following with pip:
    pip install PySide2-x.y.z.wheel

Ultimately, Qt will make them available at https://download.qt.io such
that they can be installed with:
    pip install PySide2 --index-url=https://download.qt.io/.....

However, first, Qt Releasing would like them to be tested.

To discover all of the various cherry-picks of these two patches, log
into gerrit (https://codereview.qt.io/) and search for the Change-Ids:
    I374ada7fc207d86a062f950751503764a5e5dddf
    I2dee3da70fe855859b3917e6ac775b8c5cdf5da7

There are also shiboken2-generator wheels, as well as wheels for Python
2.7 and win32 if those are needed. I did not include them here, as I did
not think they were needed.