Remove customization of Default style
The Default style block in qtquickcontrols2.conf is modifying the style
called "Default" - it's not a fallback as I was expecting - it's just
modifying properties of the style called "Default".

Since we have a preferred style and a fallback style already selected
(SwiftNav and Material, respectively), we do not need to specify
anything about the Default style, so the customizations for it are being
removed.