Drive TrackingSkyPlotTab updates from incoming data instead of timer
This drives the update of the TrackingSkyPlotTab from the incoming data
rather than a 5hz timer.

This also migrates to registering the python classes with QML via the
new @QmlElement attribute introduced in PySide6.

NOTE: This is incomplete, because the labels flicker due to deletion
after a set timeout. This logic needs to change, now that we do not
update at a predictable interval.