Add cprofile-run makefile target to run under cProfile
Add another profiling makefile target - Python's cProfile.

This aids in profiling the python code.

Usage:
	cargo make cprofile-run <toolbox args>

Once the run is complete, profiling information will be dumped to stdout.
Redirect this stdout, and analyze with a text editor.

The following is a useful post processing command line to sort on a different column:

	$ function middle()
	{
	    startidx=$1
	    len=$2
	    endidx=$(($startidx+$len))
	    filename=$3

	    awk "FNR>=${startidx} && FNR<=${endidx} { print \$0 }; FNR>${endidx} { exit }" $filename
	}
	$ (head -26 cProfile.log && middle 26 2502 cProfile.log | sort -k2) > cProfile-logs/observation-tab/cProfile.tottime-sort.log