Data Driven QML updates for LogPanel and TrackingSignals tabs (#780)
* Drive LogPanel updates from incoming data instead of timer

This drives the update of the LogPanel from the incoming data rather
than a 5hz timer.

This also migrates to registering the python logging classes with QML
via the new @QmlElement attribute introduced in PySide6.

* Remove unused emptySeries in TrackingSignalsTab

emptySeries was never used in TrackingSignalsTab, removing it.

* Drive TrackingSignalsTab updates from incoming data instead of timer

This drives the update of the TrackingSignalsTab from the incoming data
rather than a 5hz timer.

This also migrates to registering the python classes with QML via the
new @QmlElement attribute introduced in PySide6.

* Drive TrackingSkyPlotTab updates from incoming data instead of timer

This drives the update of the TrackingSkyPlotTab from the incoming data
rather than a 5hz timer.

This also migrates to registering the python classes with QML via the
new @QmlElement attribute introduced in PySide6.

NOTE: This is incomplete, because the labels flicker due to deletion
after a set timeout. This logic needs to change, now that we do not
update at a predictable interval.

* Revert "Drive TrackingSkyPlotTab updates from incoming data instead of timer"

This reverts commit 949bd2aaf5c9174e627174e43a91450444548912.

This is not yet ready, as it creates flickering due to the way that the
SkyPlot point labels are deleted and re-created. They depended on the
timer-based update interval, and now that it is at an unpredictable
update interval, another means of achieving the label updates needs to
occur.