Partial removal of 'SwiftNav' Qt Quick Control style
Because QTBUG-108518 will not be resolved in time for 6.5, we must
remove the use of custom Qt Quick Control Styles that require fallback
to another style, as the fallback style's theme does not properly
initialize. (See bug).

This removes SwiftNav style controls for all but TabBar and TabButton.

Some additional things done here:
* Remove use of SwiftButton where the customization done in SwiftButton
was not used, and Material button is more appropriate to use with the
same look and feel.
* Remove use of SwiftComboBox where the customization done in
SwiftComboBox was not used, and Material ComboBox is more appropriate to
use with the same look and feel.